Refactor DataTable components to add "className" prop
This change refactors the DataTable components in the teacher dashboard
to add a "className" prop to the TableCell component. This allows custom
styling to be applied to specific cells in the table. The "px-8" class
is added to improve cell spacing. This change improves the flexibility
and customization options for table styling. No issues are associated
with this commit.