Refactor uploadFile function and import getFileExtension
This commit refactors the `uploadFile` function in the `lib/gcp/gcp-utils.ts` file. It imports the `getFileExtension` function from the `lib/files/index.ts` file and uses it to determine the content type of the file being uploaded. This change improves the accuracy of the content type and ensures that the correct MIME type is used.